feat/deleted components and little adjustments
