--- src/components/TimeLine/TimelinePostList.component.jsx
@@ -1,12 +1,0 @@
-import React from "react"
-import TimelinePostItem from "./TimelinePostItem.component"
-
-function TimelinePosts() {
-  return (
-    <div>
-      <TimelinePostItem />
-    </div>
-  )
-}
-
-export default TimelinePosts
